Moved some methods to Effects class
